Generate type and (un)marshal functions in a single file in Go
