--- testdata/jennies/rawtypes/package-with-dashes/GoJSONMarshalling/withdashes/types_json_marshalling_gen.go
@@ -1,43 +1,0 @@
-package withdashes
-func (resource StringOrBool) MarshalJSON() ([]byte, error) {
-	if resource.String != nil {
-		return json.Marshal(resource.String)
-	}
-
-	if resource.Bool != nil {
-		return json.Marshal(resource.Bool)
-	}
-
-	return nil, fmt.Errorf("no value for disjunction of scalars")
-}
-
-func (resource *StringOrBool) UnmarshalJSON(raw []byte) error {
-	if raw == nil {
-		return nil
-	}
-
-	var errList []error
-
-	// String
-	var String string
-	if err := json.Unmarshal(raw, &String); err != nil {
-		errList = append(errList, err)
-		resource.String = nil
-	} else {
-		resource.String = &String
-		return nil
-	}
-
-	// Bool
-	var Bool bool
-	if err := json.Unmarshal(raw, &Bool); err != nil {
-		errList = append(errList, err)
-		resource.Bool = nil
-	} else {
-		resource.Bool = &Bool
-		return nil
-	}
-
-	return errors.Join(errList...)
-}
-
