--- testdata/jennies/rawtypes/variant_dataquery/GoJSONMarshalling/variant_dataquery/types_json_marshalling_gen.go
@@ -1,20 +1,0 @@
-package variant_dataquery
-import (
-	cogvariants "github.com/grafana/cog/generated/cog/variants"
-)
-
-func VariantConfig() cogvariants.DataqueryConfig {
-	return cogvariants.DataqueryConfig{
-		Identifier: "prometheus",
-	    DataqueryUnmarshaler: func (raw []byte) (cogvariants.Dataquery, error) {
-            dataquery := Query{}
-
-            if err := json.Unmarshal(raw, &dataquery); err != nil {
-                return nil, err
-            }
-
-            return dataquery, nil
-       },
-	}
-}
-
