--- testdata/jennies/rawtypes/variant_panelcfg_full/GoJSONMarshalling/variant_panelcfg_full/types_json_marshalling_gen.go
@@ -1,29 +1,0 @@
-package variant_panelcfg_full
-import (
-	cogvariants "github.com/grafana/cog/generated/cog/variants"
-)
-
-func VariantConfig() cogvariants.PanelcfgConfig {
-	return cogvariants.PanelcfgConfig{
-		Identifier: "timeseries",
-		OptionsUnmarshaler: func (raw []byte) (any, error) {
-			options := Options{}
-
-			if err := json.Unmarshal(raw, &options); err != nil {
-				return nil, err
-			}
-
-			return options, nil
-		},
-		FieldConfigUnmarshaler: func (raw []byte) (any, error) {
-			fieldConfig := FieldConfig{}
-
-			if err := json.Unmarshal(raw, &fieldConfig); err != nil {
-				return nil, err
-			}
-
-			return fieldConfig, nil
-		},
-	}
-}
-
