--- testdata/jennies/rawtypes/variant_panelcfg_only_options/GoJSONMarshalling/variant_panelcfg_only_options/types_json_marshalling_gen.go
@@ -1,20 +1,0 @@
-package variant_panelcfg_only_options
-import (
-	cogvariants "github.com/grafana/cog/generated/cog/variants"
-)
-
-func VariantConfig() cogvariants.PanelcfgConfig {
-	return cogvariants.PanelcfgConfig{
-		Identifier: "text",
-		OptionsUnmarshaler: func (raw []byte) (any, error) {
-			options := Options{}
-
-			if err := json.Unmarshal(raw, &options); err != nil {
-				return nil, err
-			}
-
-			return options, nil
-		},
-	}
-}
-
