+++ testdata/jennies/rawtypes/variant_panelcfg_only_options/GoRawTypes/variant_panelcfg_only_options/types_gen.go
@@ -1,6 +1,25 @@
 package variant_panelcfg_only_options
 
+import (
+	cogvariants "github.com/grafana/cog/generated/cog/variants"
+)
+
 type Options struct {
 	Content string `json:"content"`
+}
+
+func VariantConfig() cogvariants.PanelcfgConfig {
+	return cogvariants.PanelcfgConfig{
+		Identifier: "text",
+		OptionsUnmarshaler: func (raw []byte) (any, error) {
+			options := Options{}
+
+			if err := json.Unmarshal(raw, &options); err != nil {
+				return nil, err
+			}
+
+			return options, nil
+		},
+	}
 }
